+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Twigpack Changelog
 
+## 1.0.1 - 2018-09-22
+### Added
+* Better error logging if the manifest file can't be found (check `storage/logs/web.log`)
+* Throw a `NotFoundHttpException` if the `manifest.json` cannot be found
+
 ## 1.0.0 - 2018-09-21
 ### Added
-- Initial release
+* Initial release